VIT/Swin models refs/ symbolic types are expanded
