+++ nncf/experimental/torch/fx/nncf_graph_builder.py
@@ -32,6 +32,8 @@ class GraphConverter:
     Builds the NNCFGraph from an torch.fx.GraphModule instance.
     """
 
+    TORCH_SYMBOLIC_TYPES = (torch.SymInt, torch.SymFloat, torch.SymBool)
+
     def _get_layer_attributes(
         node: torch.fx.Node, metatype: om.OperatorMetatype, model: torch.fx.GraphModule
     ) -> BaseLayerAttributes:
@@ -196,8 +198,10 @@ class GraphConverter:
             else:
                 tensor = source_node.meta["val"]
             if isinstance(tensor, torch.Tensor):
-                tensor_shape = tuple(-1 if isinstance(i, torch.SymInt) else i for i in tensor.shape)
-            elif isinstance(tensor, torch.SymInt):
+                tensor_shape = tuple(
+                    -1 if isinstance(i, GraphConverter.TORCH_SYMBOLIC_TYPES) else i for i in tensor.shape
+                )
+            elif isinstance(tensor, GraphConverter.TORCH_SYMBOLIC_TYPES):
                 tensor_shape = (-1,)
 
         if tensor_shape is None:
