+++ nncf/quantization/algorithms/min_max/onnx_backend.py
@@ -217,7 +217,7 @@ class ONNXMinMaxAlgoBackend(MinMaxAlgoBackend):
     def get_ignored_names_by_layer_attributes(nncf_graph: NNCFGraph) -> Set[str]:
         return set()
 
-    def get_weight_nodes(self, nncf_grpah: NNCFGraph, inference_nncf_graph: NNCFGraph) -> List[NNCFNode]:
+    def get_weight_nodes(self, nncf_graph: NNCFGraph, inference_nncf_graph: NNCFGraph) -> List[NNCFNode]:
         return [node for node in inference_nncf_graph.get_all_nodes() if node.layer_attributes.has_weight()]
 
     @staticmethod
